+++ tinyui/themes/__init__.py
@@ -1,0 +1,6 @@
+#
+#  TinyUi - Themes
+#  Copyright (C) 2025 Oost-hash
+#
+
+"""Theme loading and styling."""